+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
  - Added `vertex::TransformFeedbackSession` and `transform_feedback` to the draw parameters.
+ - Added `RenderBufferAny`. Render buffers now deref to `RenderBufferAny`.
  - Deprecated `.into_vertex_buffer_any()` in favor of `.into()`.
+ - Added `.get_total_bits()` to `texture::InternalFormat`.
  - Fixed values in uniform blocks being required to implement the `Default` trait.
  - Fixed buffer sometimes not always correctly unbinded.